+++ src/providers/sinuna/SinunaAuthorizer.ts
@@ -1,6 +1,8 @@
-import axios from "axios";
+import * as jwt from "jsonwebtoken";
+
 import { AccessDeniedException } from "../../utils/exceptions";
-import { debug, logAxiosException } from "../../utils/logging";
+import { debug } from "../../utils/logging";
+import { getPublicKey } from "../../utils/openId-JWKS";
 import { leftTrim } from "../../utils/transformers";
 
 /**
@@ -10,15 +12,18 @@ import { leftTrim } from "../../utils/transformers";
  */
 export default async function authorize(idToken: string, context: string): Promise<void> {
   try {
-    const response = await axios.get(`https://login.iam.qa.sinuna.fi/oxauth/restv1/userinfo`, {
-      headers: {
-        Authorization: `Bearer ${leftTrim(idToken, "Bearer ")}`,
-      },
-    });
+    // Decode token
+    const token = leftTrim(idToken, "Bearer ");
+    const decodedToken = jwt.decode(token, { complete: true });
 
-    debug(response.data);
+    // Validate token
+    const publicKey = await getPublicKey(decodedToken, {
+      issuer: "https://login.iam.qa.sinuna.fi",
+      openIdConfigUrl: "https://login.iam.qa.sinuna.fi/oxauth/.well-known/openid-configuration",
+    });
+    const verified = jwt.verify(token, publicKey.pem, { ignoreExpiration: false });
+    debug(verified);
   } catch (error) {
-    logAxiosException(error);
     throw new AccessDeniedException(String(error));
   }
 }
